remove MapLibre as a direct dependency
