+++ src/canvas/core/coords-to-matrix.ts
@@ -1,10 +1,11 @@
 import { Euler, Matrix4, Quaternion, Vector3 } from "three";
-import { FromLngLat } from "./core/generic-map";
+import { FromLngLat } from "./generic-map";
 
 const quat = new Quaternion();
 const euler = new Euler();
 const pos = new Vector3();
 const scale = new Vector3();
+const m4 = new Matrix4();
 
 /** calculate Matrix4 from coordinates */
 export function coordsToMatrix({ longitude, latitude, altitude, fromLngLat }: {
@@ -15,5 +16,5 @@ export function coordsToMatrix({ longitude, latitude, altitude, fromLngLat }: {
   pos.set(center.x, center.y, center.z || 0);
   scale.set(scaleUnit, -scaleUnit, scaleUnit);
   quat.setFromEuler(euler.set(-Math.PI * .5, 0, 0));
-  return new Matrix4().compose(pos, quat, scale);
+  return m4.compose(pos, quat, scale).toArray();
 }
