+++ src/canvas/core/generic-map.test.ts
@@ -1,0 +1,2 @@
+// add type tests to make sure our generic functions
+// can be satisfied by Mapbox and Maplibre
