+++ src/canvas/core/use-function.ts
