+++ src/canvas/core/use-render.ts
@@ -1,9 +1,11 @@
-import { Matrix4 } from "three";
+import { Matrix4, Matrix4Tuple } from "three";
 import { StateRef } from "./state-ref";
 import { useFunction } from "./use-function";
 
+const mx = new Matrix4();
+
 export function useRender(
-  m4: Matrix4, stateRef: StateRef, frameloop: 'always' | 'demand'
+  m4: Matrix4Tuple, stateRef: StateRef, frameloop: 'always' | 'demand'
 ) {
 
   const render = useFunction((_gl: WebGL2RenderingContext, matrix: number[]) => {
@@ -11,7 +13,7 @@ export function useRender(
     const camera = stateRef.current.state.camera;
     const gl = stateRef.current.state.gl;
     const advance = stateRef.current.state.advance;
-    camera.projectionMatrix.fromArray(matrix).multiply(m4);
+    camera.projectionMatrix.fromArray(matrix).multiply(mx.fromArray(m4));
     camera.projectionMatrixInverse.copy(camera.projectionMatrix).invert();
     gl.resetState();
     advance(Date.now() * 0.001, true);
